requirements.txt
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
pytz-deprecation-shim==0.1.0.post0
tzdata==2021.5
tzlocal==4.1
